pyproject.toml
[project]
name = "augmentations_images"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "bclwn",email = "bomjtawerok777@mail.ru"},
]
dependencies = [
    "pandas>=2.2.2",
    "numpy>=1.26.4",
    "pillow>=10.3.0",
    "matplotlib>=3.9.0",
    "imgaug>=0.4.0",
    "streamlit>=1.35.0",
    "pre-commit>=3.7.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = true

[tool.pdm.scripts]
test = "pytest ."
lint = "ruff check ."
app = "streamlit run app.py"

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B"]
ignore = ["E501"]
unfixable = ["B"]

[tool.ruff.format]
quote-style = "single"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
